Move offline from dnf5 to libdnf5
We will need to share the code also with dnf5daemon-server to support
offline upgrades using D-Bus API.
