+++ dnf5/commands/system-upgrade/system-upgrade.hpp
@@ -21,7 +21,7 @@ along with libdnf.  If not, see <https://www.gnu.org/licenses/>.
 #define DNF5_COMMANDS_SYSTEM_UPGRADE_HPP
 
 #include <dnf5/context.hpp>
-#include <dnf5/offline.hpp>
+#include <libdnf5/transaction/offline.hpp>
 
 namespace dnf5 {
 
@@ -43,7 +43,7 @@ public:
 
 private:
     libdnf5::OptionBool * no_downgrade{nullptr};
-    std::filesystem::path datadir{dnf5::offline::DEFAULT_DATADIR};
+    std::filesystem::path datadir{libdnf5::offline::DEFAULT_DATADIR};
     std::string target_releasever;
     std::string system_releasever;
 };
